Various smaller fixes and cleanups

+++ src/version.hpp
@@ -10,10 +10,12 @@
  * For a full list of authors see the git log.
  */
 
+#include <cstdint>
+
 char const *get_build_type() noexcept;
 char const *get_osm2pgsql_version() noexcept;
 char const *get_osm2pgsql_short_version() noexcept;
 char const *get_minimum_postgresql_server_version() noexcept;
-unsigned long get_minimum_postgresql_server_version_num() noexcept;
+uint32_t get_minimum_postgresql_server_version_num() noexcept;
 
 #endif // OSM2PGSQL_VERSION_HPP
